+++ geoandino/locale/es/LC_MESSAGES/djangojs.po
@@ -91,3 +91,4 @@ msgstr ""
 #: geonode/static/geonode/js/upload/FileTypes.js:39
 msgid "Zip Archive"
 msgstr "Archivo zip"
+
