+++ geoandino/overrides/layers/views.py
@@ -1,0 +1,223 @@
+# -*- coding: utf-8 -*-
+
+import logging
+from django.core.urlresolvers import reverse
+from django.http import HttpResponse, HttpResponseRedirect
+from django.shortcuts import render_to_response
+from django.conf import settings
+from django.template import RequestContext
+from django.utils.translation import ugettext as _
+try:
+    import json
+except ImportError:
+    from django.utils import simplejson as json
+from django.forms.models import inlineformset_factory
+from django.forms.util import ErrorList
+
+from geonode.layers.forms import LayerForm, LayerUploadForm, NewLayerUploadForm, LayerAttributeForm
+from geonode.base.forms import CategoryForm
+from geonode.layers.models import Layer, Attribute, UploadSession
+from geonode.base.models import TopicCategory
+
+from geonode.people.forms import ProfileForm, PocForm
+from geonode.geoserver.helpers import ogc_server_settings
+
+if 'geonode.geoserver' in settings.INSTALLED_APPS:
+    from geonode.geoserver.helpers import _render_thumbnail
+CONTEXT_LOG_FILE = ogc_server_settings.LOG_FILE
+
+logger = logging.getLogger("geonode.layers.views")
+
+DEFAULT_SEARCH_BATCH_SIZE = 10
+MAX_SEARCH_BATCH_SIZE = 25
+GENERIC_UPLOAD_ERROR = _("There was an error while attempting to upload your data. \
+Please try again, or contact and administrator if the problem continues.")
+
+METADATA_UPLOADED_PRESERVE_ERROR = _("Note: this layer's orginal metadata was \
+populated and preserved by importing a metadata XML file. This metadata cannot be edited.")
+
+_PERMISSION_MSG_DELETE = _("You are not permitted to delete this layer")
+_PERMISSION_MSG_GENERIC = _('You do not have permissions for this layer.')
+_PERMISSION_MSG_MODIFY = _("You are not permitted to modify this layer")
+_PERMISSION_MSG_METADATA = _(
+    "You are not permitted to modify this layer's metadata")
+_PERMISSION_MSG_VIEW = _("You are not permitted to view this layer")
+
+from geonode.layers.views import _resolve_layer
+
+
+def layer_metadata(request, layername, template='layers/site_layers_metadata.html'):
+    """
+    Override original layer's metadata view for adding translations
+    """
+
+    layer = _resolve_layer(
+        request,
+        layername,
+        'base.change_resourcebase_metadata',
+        _PERMISSION_MSG_METADATA)
+    layer_attribute_set = inlineformset_factory(
+        Layer,
+        Attribute,
+        extra=0,
+        form=LayerAttributeForm,
+    )
+    topic_category = layer.category
+
+    poc = layer.poc
+    metadata_author = layer.metadata_author
+
+    if request.method == "POST":
+        if layer.metadata_uploaded_preserve:  # layer metadata cannot be edited
+            out = {
+                'success': False,
+                'errors': METADATA_UPLOADED_PRESERVE_ERROR
+            }
+            return HttpResponse(
+                json.dumps(out),
+                content_type='application/json',
+                status=400)
+
+        internationalize_fields()
+        layer_form = LayerForm(request.POST, instance=layer, prefix="resource")
+
+        attribute_form = layer_attribute_set(
+            request.POST,
+            instance=layer,
+            prefix="layer_attribute_set",
+            queryset=Attribute.objects.order_by('display_order'))
+        category_form = CategoryForm(
+            request.POST,
+            prefix="category_choice_field",
+            initial=int(
+                request.POST["category_choice_field"]) if "category_choice_field" in request.POST else None)
+
+    else:
+        internationalize_fields()
+        layer_form = LayerForm(instance=layer, prefix="resource")
+        attribute_form = layer_attribute_set(
+            instance=layer,
+            prefix="layer_attribute_set",
+            queryset=Attribute.objects.order_by('display_order'))
+        category_form = CategoryForm(
+            prefix="category_choice_field",
+            initial=topic_category.id if topic_category else None)
+
+    if request.method == "POST" and layer_form.is_valid(
+    ) and attribute_form.is_valid() and category_form.is_valid():
+        new_poc = layer_form.cleaned_data['poc']
+        new_author = layer_form.cleaned_data['metadata_author']
+
+        if new_poc is None:
+            if poc is None:
+                poc_form = ProfileForm(
+                    request.POST,
+                    prefix="poc",
+                    instance=poc)
+            else:
+                poc_form = ProfileForm(request.POST, prefix="poc")
+            if poc_form.is_valid():
+                if len(poc_form.cleaned_data['profile']) == 0:
+                    # FIXME use form.add_error in django > 1.7
+                    errors = poc_form._errors.setdefault('profile', ErrorList())
+                    errors.append(_('You must set a point of contact for this resource'))
+                    poc = None
+            if poc_form.has_changed and poc_form.is_valid():
+                new_poc = poc_form.save()
+
+        if new_author is None:
+            if metadata_author is None:
+                author_form = ProfileForm(request.POST, prefix="author",
+                                          instance=metadata_author)
+            else:
+                author_form = ProfileForm(request.POST, prefix="author")
+            if author_form.is_valid():
+                if len(author_form.cleaned_data['profile']) == 0:
+                    # FIXME use form.add_error in django > 1.7
+                    errors = author_form._errors.setdefault('profile', ErrorList())
+                    errors.append(_('You must set an author for this resource'))
+                    metadata_author = None
+            if author_form.has_changed and author_form.is_valid():
+                new_author = author_form.save()
+
+        new_category = TopicCategory.objects.get(
+            id=category_form.cleaned_data['category_choice_field'])
+
+        for form in attribute_form.cleaned_data:
+            la = Attribute.objects.get(id=int(form['id'].id))
+            la.description = form["description"]
+            la.attribute_label = form["attribute_label"]
+            la.visible = form["visible"]
+            la.display_order = form["display_order"]
+            la.save()
+
+        if new_poc is not None and new_author is not None:
+            new_keywords = [x.strip() for x in layer_form.cleaned_data['keywords']]
+            layer.keywords.clear()
+            layer.keywords.add(*new_keywords)
+            the_layer = layer_form.save()
+            up_sessions = UploadSession.objects.filter(layer=the_layer.id)
+            if up_sessions.count() > 0 and up_sessions[0].user != the_layer.owner:
+                up_sessions.update(user=the_layer.owner)
+            the_layer.poc = new_poc
+            the_layer.metadata_author = new_author
+            Layer.objects.filter(id=the_layer.id).update(
+                category=new_category
+                )
+
+            if getattr(settings, 'SLACK_ENABLED', False):
+                try:
+                    from geonode.contrib.slack.utils import build_slack_message_layer, send_slack_messages
+                    send_slack_messages(build_slack_message_layer("layer_edit", the_layer))
+                except:
+                    print "Could not send slack message."
+
+            return HttpResponseRedirect(
+                reverse(
+                    'layer_detail',
+                    args=(
+                        layer.service_typename,
+                    )))
+
+    if poc is not None:
+        layer_form.fields['poc'].initial = poc.id
+        poc_form = ProfileForm(prefix="poc")
+        poc_form.hidden = True
+    else:
+        poc_form = ProfileForm(prefix="poc")
+        poc_form.hidden = False
+
+    if metadata_author is not None:
+        layer_form.fields['metadata_author'].initial = metadata_author.id
+        author_form = ProfileForm(prefix="author")
+        author_form.hidden = True
+    else:
+        author_form = ProfileForm(prefix="author")
+        author_form.hidden = False
+
+    return render_to_response(template, RequestContext(request, {
+        "layer": layer,
+        "layer_form": layer_form,
+        "poc_form": poc_form,
+        "author_form": author_form,
+        "attribute_form": attribute_form,
+        "category_form": category_form,
+}))
+
+
+def internationalize_fields():
+    LayerForm.base_fields['elevation_regex'].label = _('Elevation regex')
+    LayerForm.base_fields['time_regex'].label = _('Time regex')
+    LayerForm.base_fields['metadata_uploaded_preserve'].label = _('Metadata uploaded preserve')
+    LayerForm.base_fields['is_mosaic'].label = _('Is mosaic')
+    LayerForm.base_fields['has_time'].label = _('Has time')
+    LayerForm.base_fields['has_elevation'].label = _('Has elevation')
+    LayerForm.base_fields['thumbnail_url'].label = _('Thumbnail url')
+
+
+def layer_metadata_detail(request, layername, template='layers/site_metadata_detail.html'):
+    layer = _resolve_layer(request, layername, 'view_resourcebase', _PERMISSION_MSG_METADATA)
+    return render_to_response(template, RequestContext(request, {
+        "resource": layer,
+        'SITEURL': settings.SITEURL[:-1]
+}))
