+++ geoandino/templates/layers/site_metadata_detail.html
@@ -1,0 +1,1 @@
+{% extends "overrides/geonode/metadata_details.html" %}
